+++ aiapy/psf/__init__.py
@@ -5,3 +5,4 @@ point spread function (PSF).
 
 from .deconvolve import *
 from .psf import *
+from .utils import *
