+++ changelog/367.breaking.rst
@@ -1,0 +1,1 @@
+Moved ``aiapy.psf.psf.filter_mesh_parameters`` to `aiapy.psf.filter_mesh_parameters`.
